+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- [2025-07-25] [getSchema pattern -- some TS tech debt](https://github.com/RubricLab/blocks/commit/c05ed709b751cb19916dc9c4174009794f18f358)
 - [2025-07-18] [temp fix to generic blocks](https://github.com/RubricLab/blocks/commit/9cbac5cfa1bf453210262613d0eb8fa03d624d32)
 - [2025-07-18] [cleanup](https://github.com/RubricLab/blocks/commit/499f1e2936d0515e4dd53f116b9db2841321bf33)
 - [2025-07-18] [stateful generics push](https://github.com/RubricLab/blocks/commit/bb7ae9836da634cb48896cd1221f550db609594e)